pyproject.toml
[project]
name = "master-project-deployer"
version = "0.1.0"
description = "A project for deploying master projects using Python"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
# readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pdfminer-six (>=20250327,<20250328)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
